backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ryan Loh <ryanloh29@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mem0ai = "^0.1.21"
uvicorn = "^0.32.0"
fastapi = "^0.115.2"
python-dotenv = "^1.0.1"
supabase = "^2.9.1"
groq = "^0.11.0"
requests = "^2.32.3"
langchain-core = "^0.3.12"
langchain-community = "^0.3.3"
langchain-openai = "^0.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
